chore : 스토리북 폴더 구조 변경, TODO 추가
