+++ src/components/Accordion/index.ts
@@ -1,3 +1,7 @@
 export { default as Accordion } from './Accordion';
 export { default as AccordionHeader } from './AccordionHeader';
 export { default as AccordionPanel } from './AccordionPanel';
+
+export type { AccordionType } from './Accordion';
+export type { AccordionHeaderType } from './AccordionHeader';
+export type { AccordionPanelType } from './AccordionPanel';
