+++ src/components/Pagination/PageIconButton.tsx
@@ -1,14 +1,15 @@
 /** @jsxImportSource @emotion/react */
-import React, { InputHTMLAttributes } from 'react';
+import { InputHTMLAttributes, ReactNode } from 'react';
 
 import { css } from '@emotion/react';
 
 type PageButtonType = InputHTMLAttributes<HTMLButtonElement> & {
     isDisabled: boolean;
-    icon: React.ReactNode;
+    icon: ReactNode;
     onClick: () => void;
 };
 
+// TODO: styles.ts로 파일 분리
 const iconWrapper = css({
     cursor: 'pointer',
     width: 32,
@@ -21,14 +22,14 @@ const iconWrapper = css({
     '& svg': {
         width: '16px',
         height: '16px',
-        stroke: '#8D8D8D'
+        stroke: '#8D8D8D' // TODO: palette 색상으로 변경
     },
     '&:disabled svg': {
-        stroke: '#8D8D8D4D'
+        stroke: '#8D8D8D4D' // TODO: palette 색상으로 변경
     },
     '&:hover': {
         borderRadius: 4,
-        backgroundColor: '#3378FF4D'
+        backgroundColor: '#3378FF4D' // TODO: palette 색상으로 변경
     }
 });
