+++ src/components/Pagination/index.ts
@@ -1,0 +1,3 @@
+import Pagination from './Pagination';
+
+export default Pagination;
